Plotter lables

+++ CSC1003-LinearRegression/src/plotter.h
@@ -2,7 +2,7 @@
 #ifndef PLOTTER_H
 #define PLOTTER_H
 
-extern void plotter_init(int rows, int colums, int xStart, unsigned int xLength, int yStart, unsigned int yLength);
+extern void plotter_init(int rows, int colums, float xStart, float xLength, float yStart, float yLength);
 extern void plotter_render();
 extern void plotter_dispose();
